complete property in op .def. add ut for attribute and property test

+++ framework/tests/ut/interface/src/ir/test_ir_builder.cpp
@@ -143,7 +143,7 @@ TEST(IRTEST, TestControlFlow) {
     auto fs = builder.CreateForStmt(ctx, i, constant0, batch, constant1);
 
     // test for attribute
-    fs->SetAttr("unroll", 4);
+    fs->SetAttr("unroll", int64_t(4));
 
     builder.EnterForBody(ctx, fs);
